+++ test/basic_server.py
@@ -1,6 +1,9 @@
 import argparse
+import threading
+import socket
 
 import Pyro4
+import Pyro4.naming
 
 Pyro4.config.SERIALIZER = 'json'
 
@@ -57,23 +60,37 @@ def parse_args(init_description):
     parser = argparse.ArgumentParser(description=init_description)
 
     parser.add_argument(
-        "--ns_host", "-nsn", dest='ns_host', action='store', default='localhost',
+        "--ns_host", "-nsn", dest='ns_host',
+        action='store', default='localhost',
         help="Specify a host name for the Pyro name server. Default is localhost")
 
     parser.add_argument(
-        "--ns_port", "-nsp", dest='ns_port', action='store', default=9090, type=int,
+        "--ns_port", "-nsp", dest='ns_port',
+        action='store', default=9091, type=int,
         help="Specify a port number for the Pyro name server. Default is 9090.")
 
     return parser.parse_args()
 
 
+def startNSloop(*args):
+    try:
+        return Pyro4.naming.startNSloop(*args)
+    except socket.error as err:
+        pass
+
+
 if __name__ == '__main__':
     parsed = parse_args("Start a basic server")
     bs = BasicServer()
-    # print("Registering basic server on name server")
-    # print("Server registered at {}:{}".format(parsed.ns_host, parsed.ns_port))
+
+    ns_thread = threading.Thread(
+        target=startNSloop, args=(parsed.ns_host, parsed.ns_port))
+    ns_thread.daemon = True
+    ns_thread.start()
+
     with Pyro4.Daemon(host='localhost', port=50001) as daemon:
         server_uri = daemon.register(bs, objectId='BasicServer')
         with Pyro4.locateNS(port=parsed.ns_port, host=parsed.ns_host) as ns:
             ns.register('BasicServer', server_uri)
+        print("Firing up daemon")
         daemon.requestLoop()