+++ test/helper.js
@@ -1,3 +1,5 @@
+const { spawn } = require("child_process")
+
 const { config } = require("./../lib/configuration.js")
 
 const { PromiseSocket } = require("./../lib/promise-socket.js")
@@ -6,3 +8,13 @@ config.logLevel.Proxy = "debug"
 config.logLevel.Message = "debug"
 config.logLevel.Daemon = "debug"
 config.using(PromiseSocket)
+
+const spawnPythonBasicServer = async ()=>{
+    var pythonProcess = spawn("python", ["./basic_server.py"])
+    return new Promise((resolve, reject)=>{
+        pythonProcess.stdout.once("data", resolve)
+        pythonProcess.stderr.once("data", reject)
+    })
+}
+
+exports.spawnPythonBasicServer = spawnPythonBasicServer