+++ test/helper.spec.js
@@ -1,0 +1,3 @@
+const assert = require("assert")
+
+const { spawnPythonBasicServer } = require("./helper.js")